docs: Update library package markdown-magic transformations to split "header" and "footer" details (#21853)
The existence of a singular "library package README" transformation
encourages developers to not include other necessary details in package
READMEs. It was also previously unused in a lot of places because it was
difficult to interleve the common details with custom ones.

This PR removes that template and introduces 2 new ones: a "header"
template and a "footer" template, with recommended guidance for where
each should be used.

Also updates existing consumers of the removed template to migrate to
the new ones.