improvement(fluid-build): Use hash-based caching in Biome task (#21795)
We've observed that the Biome fluid-build tasks are often triggered when
they shouldn't be. The problem seems to be in the file modified time,
which doesn't match expected, implying a "rogue" process has touched a
file.

I added the ability to use a content-based hash instead of the file
stats in a previous change, so I just updated the task to use it. This
sidesteps any issues with the stats-based approach.

This mechanism is ~20% slower than the stats-based approach, but tasks
with few files/smaller files are less affected. The Biome task is most
affected because it effectively caches every file in the path. #21826
will add support for ignore files in the Biome task, which should limit
the perf hit for that task.

The Biome task was introduced without content-based caching and that is
different than the Prettier task behaved. With this change the behavior
should match the Prettier task more closely and should get us back to a
known good config.