Add e2e test vscode workspace with more flexible launch targets (#21905)
## Description

Adds a VSCode workspace set up for e2e test debugging. It has the
following advantages over the primary (root) workspace:

- Access to legacy module installation from the file explorer
- Dedicated launch configurations for each driver we test against
- Inputs selected at launch time for:
    - driver-specific config (e.g. odsp endpoint)
    - console output
- whether break points can be set in legacy modules (this is a fairly
large performance hit, so default is 'no')
- Tinylicious launch target is a compound target which allows
simultaneous server + client debugging
- FF code under `.legacy` (i.e. the modules installed for compat
testing) isn't under skipFiles by default, so break points in this code
works (either in the js if source map loading is turned off, or in the
ts if it is turned on, which is configurable at launch time)

The r11s tests are difficult to run on Windows, so I didn't test that
launch target. Other targets all work well as far as I can tell (which
includes setting breakpoints in various places in dependent packages)

Sample view of launch window when using this workspace:


![image](https://github.com/user-attachments/assets/e00569b6-9be0-4177-92f5-8e16d5587c87)

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>