tree: Improve ModularChangeFamily asymptotics (#21521)
## Description

This PR refactors ModularChangeFamily to use algorithms for rebasing and
composition which do not require inspecting irrelevant portions of the
changeset trees. Field kinds are now only required to call `rebaseChild`
or `composeChild` when both changesets have edits for that child node.

Further work is still necessary to make the cost of
ModularChangeFamily's compose and rebase scale proportional to the size
of the intersection between the input changesets, instead of scaling
proportional to the size of both changesets. Tasks not addressed in this
PR include incremental pruning of changesets, keeping aliasing paths
short in the new node alias table, and converting various maps and
arrays in ModularChangeset to data structures which can be efficiently
composed.

---------

Co-authored-by: yann-achard-MS <97201204+yann-achard-MS@users.noreply.github.com>