build(client): Bump dependencies on build-tools to 0.41.0 (#21933)
Bumps the client dependencies on build-tools to 0.41.0.

## Automated changes

- The type test generator was updated in this build-tools release, so
the output of all type tests was regenerated.

## Manual changes

- The type test generator uses new names in some cases, which means
existing test disables in package.json had to be renamed. This extra
step could have been avoided if I'd done the bump after the next
release, when the type tests are all reset and the disables are all
reset. But this release contains useful changes and is worth getting
into the release branch as well.
- The azure-client disables needed an additional entry,
"RemovedClassStatics_AzureFunctionTokenProvider" which is expected.