tree: Avoid adding an empty changeset to compositions (#21958)
## Description

This PR removes a behavior where ModularChangeFamily, as well as some
test change handlers, would include an empty changeset into the
composition whenever composing an array of changesets. This was done so
that even composing a single changeset would cause a full composition
pass. In the past this was necessary because the composition pass had
the side effect of inlining revision tags. Now that revision tag
inlining is done separately, this pass is unnecessary.