ci: Label PRs that add/modify changesets (#21954)
It's helpful when preparing for a release to check if any pending PRs
are adding changesets. If the PRs were labeled, doing that query becomes
trivial.

I updated the area labeler to add the
[changeset-present](https://github.com/microsoft/FluidFramework/issues?q=sort%3Aupdated-desc+is%3Aopen+label%3Achangeset-present)
to changes with edits to changesets. This does nothing besides add the
label. Nothing else is automated or changed.