build(client): Override AWS dependencies in lockfile (#21979)
oclif includes some AWS-related features but we don't use them, so the
AWS dependencies end up causing a lot of unneeded churn in our
lockfiles.

I used pnpm overrides to replace the AWS-related oclif dependencies with
an empty package. This should not have downstream effects since we are
not using these oclif capabilities.