refactor(markdown-magic): Simplify template names (#21964)
See the updates in `markdown-magic/README` for the full list of renames.

Also note that in addition to the rename from `EXAMPLE_PACKAGE_README`
TO `EXAMPLE_README_HEADER`, that template was updated to _only_ include
"header" contents (for now, this is just the "Getting Started" section).
It now omits the "footer" contents which are now available via the
renamed `README_FOOTER` template. There were no existing usages of this
updated template outside of tests, so no usages had to be updated. I
will likely update many of our examples to use this updated template in
a follow-up PR.