[GC] Remove deprecated getBaseGCDetails (#21990)
## Description
getBaseGCDetails and related code were deprecatedmore than a year ago by
https://github.com/microsoft/FluidFramework/pull/14196. This PR removes
it.

## Reviewer guidance
This is targetting a test branch (test/gc-2.3). Since this has legacy
alpha API breaking changes which are not allowed to be merged until
v2.3, this is a workaround to get the changes reviewed and merged in a
test branch. Later when breaking changes are allowed, these will get
merged to main.