tree: To map tree work (#21995)
## Description

Cleans input type down-casting logic.

This removes some dead code, does some refactors, and fixes a couple
issues:

1. fields with defaults did not translate the API key to the stored key.
This is not called out in the changeset since the only fields with
defaults that matter for this are identifiers, and the current API makes
it impossible to select a different stored key.
2. the types for implicit node construction of arrays and maps did not
match the runtime behavior. This is improved, tested and documented, but
not fully fixed.