Deprecate summarizeProtocolTree in ILoaderOptions (#21999)
Marked `summarizeProtocolTree` as deprecated in the extended
`ILoaderOptions` in container layer. Also created an ADO item to remove
the declaration and it's usage in the future versions when we can alter
the alpha apis.

summarizeProtocolTree property in ILoaderOptions (the extended
definition in container layer) was added to test single-commit summaries
during the initial implementation phase. The flag is no longer required
or should be used, so marked it as deprecated. If a driver needs to
enable or disable single-commit summaries, it can do so via
IDocumentServicePolicies.