Add ConsensusQueue to handle validation test suite (#22014)
## Description

Adds back `ConsensusQueue` to the handle validation tests. The test had
a few issues preventing it from working with ConsensusQueue. The direct
cause for the hang that the original bug reported is that container1 was
disposed before attempting to read the handle, and ConsensusQueue
requires ops to round-trip before its data can be read.

While working on this, I also refactored the affected test and suite a
fair bit for clarity, and tweaked what the test case was doing. I split
it into a couple different cases based on what the handle in the DDS was
referencing, as [previous
discussion](https://github.com/microsoft/FluidFramework/pull/20995#discussion_r1591574035)
on the topic mentioned the problematic case was something not being
explicitly tested.


[AB#8058](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8058)

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>