Make isFluidHandle public (#22029)
## Description

This makes isFluidHandle public.

An alternative API, using Symbol.hasInstrance and `instancof
IFluidHandle` is not possible here, since there already is non-type
value exported under the name IFluidHandle (its the string
"IFluidHandle") which can not have the symbol added to it.

This leaves the existing isFluidHandle function seeming like the only
real API option.

Detecting handles correctly requires knowing implementation details of
handle and dealing with the possibility of old in memory handle
representations, so we don't want customers rolling their own. As there
are use-cases for handle detection, we should probably expose this API
to do it robustly.