Fix sharedString legacy snapshot where there is 0 segements with attribution yet  (#22044)
## Description

Attribution is not initialized on segments yet as min seq number could
still be 0. So, while snapshotting we don't set the attribution prop on
segment. So, when we reload the segment, then in extractAttribution() in
snapshpotLoader.ts file, we detach from the attribution policy since if
the summarized segments do not have attribution, then loaded file should
not have attribution too as they are enabled on new file and then all
future loads should have attribution.

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>