(docs): use Hugo breadcrumb to fix link checker (#22063)
Current custom breadcrumb doesn't seem to work properly since it relies
on the URL string to generate the breadcrumb links. However, this
generations method seems to be failing in scenarios where a links parent
endpoint doesn't match the display name.

e.g. 
for the following endpoint: 
/docs/api/v1/container-definitions/connectionstate/catchingup-typealias/

it will assume that the parent endpoint is:
/docs/api/v1/container-definitions/connectionstate/

However the actual endpoint is:
/docs/api/v1/container-definitions/connectionstate-namespace/

Consequently, the custom breadcrumb html will generate a link with a
non-working endpoint which causes link checker to fail. Updated to use
the default Hugo breadcrumb generator instead which works similar to the
current breadcrumbs. Note there is a slight change visually with the
breadcrumb links being below the title instead of above.

before:

![docs-breadcrumb](https://github.com/user-attachments/assets/79364ec7-9684-4944-9f44-7b55ba0fb1ca)

after:

![docs-breadcrumb2](https://github.com/user-attachments/assets/dbf1b061-8829-4b1d-8948-66ab9f098131)