Add `typedJsonCursor` and fix up lazy tree tests (#22062)
## Description

This adds a small extension to the existing `singleJsonCursor` which
allows the JSON objects to be conveniently annotated with the type of
the corresponding node. This allows much of our non-primitive test
content to be inserted as a cursor without relying on the
`schemaAware`/`contextuallyTyped` libraries.

lazyField.spec.ts and lazyNode.spec.ts are updated to use
`singleJsonCursor` and `typedJsonCursor` (where necessary) for inserted
content. Other test files will follow in future PRs.