[GC] Remove used routes state from data store context (#21939)
Currently, if a data store is not realized and updateUsedRoutes() is
called on it, the used routes are saved and are passed on the child
DDSes (and their summarizer nodes) during data store realization. This
PR removes storing the used routes state from data store context. The
summarizer node should pass this information directly to the child
summarizer node for couple of reasons:
- If the realization of the data store fails after the child summarizer
node is created, the child node will not have used routes leading to
inconsistent state. This has happened in production before.
- Since the used routes along with other GC state is maintained and
tracked by the summarizer nodes, the data stores should not be saving it
as that leaks the state to different layers.


[AB#3448](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3448)