MergeTree: Fix annotate rebase properties (#22069)
There is an existing issue with how we regenerate annotates for
resubmit. The specific problem is when the segment is inserted with no
properties but is subsequently annotated. In this case the existing
behavior will insert the segment with the annotated properties, which is
incorrect. The segment should be inserted with no properties, and the
subsequently resubmitted annotate will apply the properties.

The added unit test previously failed, and now passes.

---------

Co-authored-by: Abram Sanderson <Abram.sanderson@gmail.com>