ci: Fix docker pipelines to correctly pack packages (#22072)
This change reverts part of the changes made in #21018. The past changes
inadvertently caused the packages to be published without any built
content. I have verified from test builds that the published packages do
have built content with this change.

In this change, the pack process for docker pipelines is once again run
with a unique shell command that is run in the docker container, and the
package lists are created directly in the pipeline instead of by a
script.

This is unfortunate from a maintenance perspective because it means
there are two slightly different pack paths depending on the pipeline.
That said, this is by far the most straightforward fix.