Move obliterate reconnection under separate flag and improve behavior (#22080)
The obliterate logic for reconnect is not well defined. This change puts
that behavior under a separate flag to ensure reconnects with obliterate
are only done intentionally. Secondarily, this change modifies how it
works with reconnect when there is a remote obliterate, and a local
insert. The change essentially rolls back the insert, rather than having
it bypass the obliterate.

related
[AB#3714](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3714)

---------

Co-authored-by: Abram Sanderson <Abram.sanderson@gmail.com>
Co-authored-by: Joshua Smithrud <54606601+Josmithr@users.noreply.github.com>