[GC] Cleanup configs and options that are not used or needed anymore (#21992)
Removed the following configs / features that are not used or are not
needed anymore. These were added throughout the GC feature development
to roll out certain functionalities in phases. Others were asked by our
partners and are either not used anymore or were never used.

The goal is to keep the configs simple as following:
- Mark phase - GC mark phase is always enabled and cannot be disabled.
Only very old document before GC was added will have them disabled.
- Sweep phase - Sweep phase is disabled by default. It can be enabled
via the `enableGCSweep` GC option. Requires GC mark phase to be enabled.
- Tombstone phase
- Marking objects as tombstone is always enabled and cannot be disabled.
Note that this won't cause failing tombstoned objects. That is discussed
in the next entry.
- If sweep is enabled, throwing on loading tombstone objects is enabled.
It can only be disabled via the config
"Fluid.GarbageCollection.DisableThrowOnTombstoneLoad". I don't think we
need this because even if tombstone is disabled, sweep is still enabled
and so, it might be better to throw on tombstone in order to trigger
autorecovery.

Here's the list of changes:
- Tombstone
- `Fluid.GarbageCollection.DisableTombstone` - This config allowed
disabling tracking tombstones and marking objects as tombstones. We
always want this enabled as it doesn't result in any errors or failures.
- `gcDisableThrowOnTombstoneLoad` - This GC option allowed disabling
throwing on tombstone load. This is not used and there is a config to
toggle this behavior anyway that we are keeping -
"Fluid.GarbageCollection.ThrowOnTombstoneLoadOverride".
- `Fluid.GarbageCollection.ThrowOnTombstoneUsage` - This config was
never used. It allows throwing an error when a tombstone object is used.
However, throwing on tombstone load is sufficient as we learnt from our
roll out. An object has to be loaded before it is used.
- `tombstoneEnforcementAllowed`, `throwOnTombstoneLoad` and
`throwOnTombstoneUsage` - These were exposed by GarbageCollector. The
first two were used to simply log with tombstone errors. The feature for
the last one has been removed. We don't need to log these because they
are logged when Garbage collector loads.
- `tombstoneMode` - This has been removed from GC configs. Tombstone
mode cannot be disabled anymore.
- Sweep
  - `sweepAllowed` - This is not exposed from the GC config anymore.
- `shouldRunSweep -> sweepEnabled` - Renamed `shouldRunSweep` ->
`sweepEnabled`. This is a session level setting telling whether sweep is
enabled for a session. Also, the type is changed from enum to boolean.
Now, sweep can either be turned fully on or fully off (see the next
entry).
- `disableDataStoreSweep` and
`Fluid.GarbageCollection.DisableDataStoreSweep` - These configs / GC
options were used when rolling out sweep to first roll out for
attachment blobs only. These are not needed anymore and have been
removed.
- `gcEnabled -> gcAllowed` - GC allowed is more inline for document
level settings. This setting tells whther GC is allowed on a document.
- `sendGCUnexpectedUsageEvent` - This helper function has been removed.
The only thing it added was logging various tombstone flags. Most of
these have been removed and others are logged when garbage collector
loads. The package path it logged is already logged for all data store
context events by default.


[AB#9096](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/9096)