feat(tree): Add isRepoSuperset to validate the incompatibilities between … (#22045)
…two repos


[AB#8717](https://dev.azure.com/fluidframework/internal/_workitems/edit/8717)

The PR consists of two parts:

1. Fixing previous unit tests for `allowsRepoSuperset`

The previous tests contained schemas that were incorrectly treated as
`neverTree`, it may not change the test result but influence the the
analysis process of `allowsRepoSuperset`. This fix ensures accurate
schema creation and validation.

2. Introducing `isRepoSuperset`:

This function uses incompatibilities to determine if changes to a
document schema are backward-compatible. According to the [doc of schema
evolution](https://github.com/microsoft/FluidFramework/blob/main/packages/dds/tree/docs/user-facing/schema-evolution.md),
`isRepoSuperset` supports three types of changes:

- Adding an optional field to an object node.

- Expanding the set of allowed types for a field.

- Relaxing a field kind to a more general field kind.

Notes: We expect `isRepoSuperset` to return consistent results with
`allowsRepoSuperset`. However, there are some inconsistencies:

- **Different Node Kinds**: If a and b have different node kinds (e.g.,
a is an `objectNodeSchema` and b is a `mapNodeSchema`), `isRepoSuperse`
will determine that a can never be the superset of b. In contrast,
`allowsRepoSuperset` will continue validating internal fields.


## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).

---------

Co-authored-by: Abram Sanderson <Abram.sanderson@gmail.com>