refactor(ci): Increase timeout of local server e2e tests from 3s to 15s (#22110)
Despite increasing the timeout for local server e2e tests from 2s to 3s
https://github.com/microsoft/FluidFramework/pull/21928, timeout issues
continue to occur. We determine to aggressively bump the timeout
threshold as a mitigation. Given that each previous failure involved
only one timeout, and there's a 1-hour hard cap for local server tests,
this adjustment should be safe.

In addition, we need to identify and address the root cause ASAP and
restore the timeout to 2s once the issue is resolved, tracked by
[AB#8942](https://dev.azure.com/fluidframework/internal/_workitems/edit/8942)

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).

---------

Co-authored-by: Mark Fields <markfields@users.noreply.github.com>