Remove PropertyDDS to Tree Schema Converter (#22111)
## Description

This schema converter was out of date and missing key features.
Since it deepened internal on API we want to remove, it has been
removed.
It may be restored with some updates in the future.

## Breaking Changes

This schema converter has been removed. Using it required using internal
tree types, so this should only have been usable internally, and there
were no internal users.