[GC] Removed throwOnInactiveLoad GC option and the config to disable autorecovery (#22107)
Removed the following:
- `throwOnInactiveLoad` GC container runtime option and
`throwOnInactiveLoad` internal GC config - These were for a feature to
thrown an error when an inactive data store is loaded. This was an
experimental feature added during sweep development. However, it was
never used.
- `"Fluid.GarbageCollection.DisableAutoRecovery"` feature config and
`tombstoneAutorecoveryEnabled` internal GC config - These were added
while shipping auto recovery feature to turn it off in case it caused
issues. This has been enabled for a while and we want it to always be on
if a tombstone data store is loaded.