tree: Fix bug with incremental rebasing (#22119)
## Description

Fixed a bug with rebasing over a changeset which moves a node into a
subtree which is also moved. If the rebasing change previously did not
touch that subtree but had some changes in its field, it could lose the
changes for that subtree when rebasing.

Previously it was assumed that rebasing required at most two passes over
any field: a first pass to generate information about moves, and a
second pass which could consume any incoming move information. This PR
adds a pass for fields which have changes moved into a subtree untouched
by the rebasing changeset. The node change may be discovered after the
first pass, but may also need to be included in the move information for
the field, so we essentially redo the first pass for these fields before
processing other invalidated fields.