Fix Fuzz Test Replay and Minimization with stashing (#22116)
There was a problem with the operation format of stashing that made
replay and minimization not work. Specifically, the stashClient
operation had the container runtime id, which was a random uuid, so on
replay the client id could not be found, i've changed this to the fuzz
test client id, so it remains consistent. There was a related problem
with the new clients id, so i've changed that to append an instance
number to the original client's id, such that it can be differentiated
from the original client, but still associated with it.