ci: Update release notes tracking issue (#22102)
Adds a GitHub actions workflow to update an issue with the current
release notes every time changesets are added or changed on the main
branch. This will provide a continually updated snapshot of what the
release notes will look like when we release.

I also updated some changeset content to better work with the release
note layout. There will be additional changes made in #22067.

---------

Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>
Co-authored-by: Craig Macomber (Microsoft) <42876482+CraigMacomber@users.noreply.github.com>
Co-authored-by: jzaffiro <110866475+jzaffiro@users.noreply.github.com>
Co-authored-by: Joshua Smithrud <54606601+Josmithr@users.noreply.github.com>