Fix messageSize e2e test (#22156)
The flakiness of this test was caused by the large messages being sent
and processed. Service or the client wouldn't be able to keep up fast
enough consistently to handle these large ops, thus causing this test to
timeout. By reducing the size of the ops being sent, and subsequently
the compression/chunking limits, this test should be much less flaky.


[AB#8608](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8608)