Flaky test fix: Run summarizer test only for local and odsp (#22153)
_"Closes the summarizing client instead of refreshing..."_ test was
failing because of the test timing out for FRS service. The code and the
test seem to work as expected. On logging timestamps it was found that
every api request takes > 1.3 seconds or more to run for FRS causing the
test to timeout.
The test is about confirming the summarizer behavior, and it doesn't
matter which service it runs against. With this PR, I am limiting the
service endpoints to be `local` and `odsp`(for covering the scenario of
single commit summaries as well, just as an extra caution).

[AB#10972](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/10972)