Use MapTrees as flexible node content (#22165)
## Description

This updates the flex tree layer to use MapTrees as input for new
content, rather than cursors.

### Why?

* Unhydrated nodes are soon going to be editable, and in order for that
to work they need direct access to the MapTrees that they are handed as
input. Without this change, the inserted content would need to "round
trip" from a MapTree to a cursor and then back to a MapTree, which not
only loses the identity of the MapTree object (important for caching of
unhydrated nodes' backing state) but is also unnecessarily inefficient.
* The distinction between FlexibleNodeContent and FlexibleFieldContent
is now more type safe - the former is a single object and the latter is
an array, rather than each being a cursor whose mode ("node" or "field")
cannot be evaluated at compile time.

### How?

All production codepaths from simple-tree already use MapTrees for new
content, so no significant changes were required to support this. A
smattering of tests needed to be updated by adding a conversion wrapper
for new content (though, many of those tests will be removed in the
coming days anyway as the flex layer continues to be stripped down).