Expose `editor` for flex fields (#22170)
## Description

This unifies the editing API of the flex layer by directly exposing the
field editor. Sequence fields already do this, however, they expose an
editor that uses cursors for new content. This updates the editor
exposed by sequence fields to accept MapTrees as new content, and
introduces the same pattern for value and optional fields. The flex
layer now translates from the MapTrees to cursors, rather than it being
done externally. This is part of a continued effort to simplify and
reduce the API surface of the flex layer.

This also removes the content setter from value/optional fields, since
it is now redundant with the editor.