loadContainerPaused() test fix (#22173)
**Issue:** `Throw if attempting to pause at a sequence number before the
latest summary` test was failing erratically. Passed for local server
but failed for odsp/frs etc services.

**Root Cause:** The test tries to check that the `loadContainerPaused()`
function should throw an error when we try to load from a seq number
which is lesser than the latest snapshot sequence number. This was
failing because the latest snapshot that the container refereed to was
an older one (the very first snapshot generated when a container is
created) because the`fetchSnapshot` races to retrieve the snapshot from
cache or network and in almost all cases, refers to the snapshot in
cache.

**Fix:** Now we are explicily providing the snapshot version number in
the `IRequestHeaders` so that the test refers to the latest snapshot
when performing comparisons.


[AB#8561](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8561)