server: explicitly reject requests for expired Ephemeral Containers (#22174)
## Description

Currently, we rely on an Ephemeral Container to either 1) be cleaned up
by the Deli lambda on session end, or 2) expire due to DB and Redis TTL
values.

There are inconsistencies in configurations and TTL behaviors regardless
of configs, so we want to explicitly reject access to Ephemeral
containers that are older than a certain time.

This PR causes all Historian requests and Alfred getSession requests to
fail with an explicit `404 - Ephemeral Container Expired: ...` error
when the container was created longer ago than the EphemeralDocumentTTL
config value. It also changes Gitrest's Ephemeral TTL configuration to
use an explicit EphemeralDocumentTTL value for consistency, rather than
an implicit general Redis TTL value.

The defaults for these values are remaining as 24 hours.