fix(ci): Make caching of compat version install location conditional (#22190)
## Description

Follow up to https://github.com/microsoft/FluidFramework/pull/22181

The service client e2e tests pipeline uses the same template as the
(non-service-client) e2e tests pipeline and doesn't do compat of any
kind, so it never creates the install location for compat versions and
the caching task is completing unsuccessfully due to that. This PR makes
the caching behavior opt-in through a parameter that only the e2e tests
pipeline sets to the necessary value.