Remove `move` method from sequence editor (#22199)
## Description

The `move` method on `SequenceEditor` is redundant and less powerful
than the `move` method on `IDefaultEditBuilder` which can accomplish
cross-field moves in addition to the same-field moves that the `move`
method on `SequenceEditor` can do. It is only used in tests, so this
updates the tests to use the `move` method on `IDefaultEditBuilder`
instead.